chore(infrastructure): move indexes & collection
