+++ src/config/mongodb.ts
@@ -1,5 +1,5 @@
+import { workspacesCollectionName } from '@/kairos/shared/infrastructure/mongodb/schema/workspaces-collection-schema';
 import { assertIsNonEmptyString } from '@/kairos/shared/kernel/assert';
-import { workspacesCollectionName } from '@/kairos/shared/persistence/mongodb/workspaces-collection-schema';
 
 const connectionString = process.env.MONGODB_CONNECTION_STRING;
 assertIsNonEmptyString(connectionString, 'MONGODB_CONNECTION_STRING is not defined in environment variables');
