+++ src/kairos/setup/infrastructure/repository/exists-super-admin.ts
@@ -1,5 +1,5 @@
 import { type ExistsSuperAdmin } from '@/kairos/setup/domain/super-admin-repository';
-import { type UsersCollection } from '@/kairos/shared/persistence/mongodb/users-collection-schema';
+import { type UsersCollection } from '@/kairos/shared/infrastructure/mongodb/schema/users-collection-schema';
 
 type ExistsSuperAdminDependencies = Readonly<{
   usersCollection: UsersCollection;
