+++ src/kairos/setup/setup-composition.ts
@@ -11,7 +11,7 @@ import { createCreateSuperAdmin } from '@/kairos/setup/infrastructure/repository
 import { createExistsSuperAdmin } from '@/kairos/setup/infrastructure/repository/exists-super-admin';
 import { mongoDBClient } from '@/kairos/shared/infrastructure/mongodb/client/client';
 import { usersCollection } from '@/kairos/shared/infrastructure/mongodb/collection/users-collection';
-import { ensureMongoIndexes } from '@/kairos/shared/persistence/mongodb/ensure-mongo-indexes';
+import { createEnsureIndexes } from '@/kairos/shared/infrastructure/mongodb/index/create-ensure-indexes';
 import { hashPassword } from '@/kairos/shared/security/password';
 
 const existsSuperAdmin: ExistsSuperAdmin = createExistsSuperAdmin({ usersCollection });
@@ -22,7 +22,7 @@ const createSuperAdminSetupTask = createEnsureSuperAdminTask({ existsSuperAdmin,
 function createRuntimeInfrastructureTasks(): RuntimeInfrastructureTask[] {
   return [
     mongoDBClient.connect.bind(mongoDBClient),
-    ensureMongoIndexes({ database: mongoDBClient.db(), indexes: mongodbConfig.indexes }),
+    createEnsureIndexes({ database: mongoDBClient.db(), indexes: mongodbConfig.indexes }),
   ];
 }
