+++ src/kairos/shared/infrastructure/mongodb/collection/workspaces-collection.ts
@@ -2,6 +2,6 @@ import { mongoDBClient } from '@/kairos/shared/infrastructure/mongodb/client/cli
 import {
   WorkspacesCollection,
   workspacesCollectionName,
-} from '@/kairos/shared/persistence/mongodb/workspaces-collection-schema';
+} from '@/kairos/shared/infrastructure/mongodb/schema/workspaces-collection-schema';
 
 export const workspacesCollection: WorkspacesCollection = mongoDBClient.db().collection(workspacesCollectionName);
