+++ src/kairos/shared/infrastructure/mongodb/schema/users-collection-schema.ts
