+++ src/kairos/shared/infrastructure/mongodb/schema/workspaces-collection-schema.ts
