+++ tests/__vitest__/integration-test.ts
@@ -1,12 +1,12 @@
 import { mongodbConfig } from '@/config/mongodb';
 import { mongoDBClient } from '@/kairos/shared/infrastructure/mongodb/client/client';
-import { ensureMongoIndexes } from '@/kairos/shared/persistence/mongodb/ensure-mongo-indexes';
+import { createEnsureIndexes } from '@/kairos/shared/infrastructure/mongodb/index/create-ensure-indexes';
 import { afterEach, beforeEach } from 'vitest';
 
 export function integrationTest(): void {
   beforeEach(async () => {
     await mongoDBClient.connect();
-    await ensureMongoIndexes({
+    await createEnsureIndexes({
       database: mongoDBClient.db(),
       indexes: mongodbConfig.indexes,
     })();
